refactor: 크레딧에 대한 로직을 CreditService로 이동 (#187)
