+++ src/main/java/com/newfit/reservation/domains/reservation/controller/ReservationApiController.java
@@ -4,6 +4,7 @@ import com.newfit.reservation.common.auth.AuthorityCheckService;
 import com.newfit.reservation.domains.authority.domain.Authority;
 import com.newfit.reservation.domains.authority.service.AuthorityService;
 import com.newfit.reservation.domains.credit.dto.request.ObtainCreditRequest;
+import com.newfit.reservation.domains.credit.service.CreditService;
 import com.newfit.reservation.domains.reservation.domain.Reservation;
 import com.newfit.reservation.domains.reservation.dto.request.ReservationRequest;
 import com.newfit.reservation.domains.reservation.dto.request.ReservationUpdateRequest;
@@ -28,9 +29,10 @@ import static org.springframework.http.HttpStatus.CREATED;
 @RequiredArgsConstructor
 public class ReservationApiController {
 
-    private final ReservationService reservationService;
     private final AuthorityService authorityService;
     private final AuthorityCheckService authorityCheckService;
+    private final CreditService creditService;
+    private final ReservationService reservationService;
 
     @PostMapping
     public ResponseEntity<Void> reserve(Authentication authentication,
@@ -108,7 +110,7 @@ public class ReservationApiController {
         Authority authority = authorityService.findById(authorityId);
 
         reservationService.updateStatusAndCondition(reservation);
-        reservationService.checkConditionAndAddCredit(reservation, authority, request.getEndEquipmentUseAt());
+        creditService.checkConditionAndAddCredit(reservation, authority, request.getEndEquipmentUseAt());
 
         return ResponseEntity
                 .noContent()
